Add files via upload
the html script is unable to open the vega file with the following suffixes: ".vega" and ".js.vega". 

The HTML must instead contain ".js.json".

Not sure why this is the case, it must have something to do with the file types and html not being able to open the other suffixes?